requirements.txt
torch
torchvision
transformers
evaluate
jiwer
numpy

peft
sentencepiece
bitsandbytes
accelerate

opencv-python
pillow
fontTools

pyyaml
protobuf
duckdb